+++ packages/jazz-browser/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies
-  - jazz-js@0.7.0-alpha.0
+  - jazz-tools@0.7.0-alpha.0
   - cojson@0.7.0-alpha.0
   - cojson-storage-indexeddb@0.6.4-alpha.0