+++ packages/jazz-nodejs/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Patch Changes
 
 - Updated dependencies
-  - jazz-js@0.7.0-alpha.0
+  - jazz-tools@0.7.0-alpha.0
   - cojson@0.7.0-alpha.0
   - cojson-transport-nodejs-ws@0.5.2-alpha.0