+++ packages/jazz-react/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 - Updated dependencies
   - jazz-browser@0.7.0-alpha.0
-  - jazz-js@0.7.0-alpha.0
+  - jazz-tools@0.7.0-alpha.0
   - cojson@0.7.0-alpha.0
 
 ## 0.5.4